pyproject.toml
[tool.poetry]
name = "fluid-sim"
version = "0.1.0"
description = ""
authors = ["Philip Youssef <pyoussef@rippling.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.0.0"
pygame = "^2.6.0"
pyopengl = "^3.1.7"
moderngl = "^5.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
